Add files via upload
Th project is based on the Markov Models of Natural Language which when seeded with a small amount of text, create a large corpus of text similar to the seed text. Here I use first 10 chapters of Jane Austins book "Emma" to create a much simpler language model to generate text. I create an nth - order Markov Model with creates a string of text one letter at a time based on the most recent n block of letters. The model then adds the last letter of the n+1 block of text based on its probability of occurrence in the entire text( first 10 chapters of Jane Austins book "Emma").